requirements.txt
scipy>=1.9.0
numpy>=1.23.2
pandas>=1.4.3
progressbar2==4.0
pyyaml>=6.0
pytest>=7.1.2
